added test for Delete /todos/:id route

+++ server/server.js
@@ -48,9 +48,9 @@ app.get('/todos/:id', (req, res) => {
 
 app.delete('/todos/:id', (req, res) => {
     var id = req.params.id;
-    // if(!ObjectID.isValid(id)) {
-    //     return res.status(404).send();
-    // };
+    if(!ObjectID.isValid(id)) {
+        return res.status(404).send();
+    };
 
     Todo.findByIdAndRemove(id).then((todo) => {
         if(!todo) {
